ai-worker/services/image_processor.py
"""Image processing service with AI model integration."""

import logging
from typing import Tuple
from PIL import Image
from io import BytesIO


class ImageProcessor:
    """
    Handles AI-based image processing for anime style transformation.
    
    TODO: Integrate actual AI model
    Currently returns the input image as a placeholder.
    
    """

    def __init__(self, logger: logging.Logger):
        """
        Initialize the image processor.
        
        Args:
            logger: Structured logger instance
        """
        self.logger = logger
        self.logger.info("ImageProcessor initialized (placeholder mode)")

    async def process_image(self, job_id: str, image_data: bytes) -> Tuple[bytes, str]:
        """
        Process an image to transform it into anime style.
        
        Args:
            job_id: Unique job identifier for logging
            image_data: Input image as bytes
            
        Returns:
            Tuple of (processed image bytes, content type)
            
        TODO: Replace placeholder with actual AI model
        """
        try:
            # Open the image to validate it
            image = Image.open(BytesIO(image_data))
            self.logger.info(
                "Processing image (placeholder)",
                extra={
                    "job_id": job_id,
                    "format": image.format,
                    "size": f"{image.width}x{image.height}",
                    "mode": image.mode,
                },
            )

            # TODO: AI model processing goes here
            output_buffer = BytesIO()
            
            image.save(output_buffer, format="PNG")

            output_data = output_buffer.getvalue()

            self.logger.info(
                "Image processed successfully (placeholder)",
                extra={"job_id": job_id, "output_size": len(output_data)},
            )

            return output_data, "image/png"

        except Exception as e:
            self.logger.error(
                f"Error processing image: {e}",
                extra={"job_id": job_id, "error": str(e)},
            )
            raise
