ai-worker/pyproject.toml
[project]
name = "ai-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.128.5",
    "minio>=7.2.20",
    "nats-py>=2.13.1",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
